+++ src/commands/react.js
@@ -56,7 +56,6 @@ Commands:
       ${opt('--mount-id')}     id for the <div> the app will render into ${opt('[default: app]')}
       ${opt('--no-fallback')}  disable serving of the index page from any path
       ${opt('--no-hmr')}       disable use of Fast Refresh for Hot Module Replacement
-      ${opt('--no-polyfill')}  disable inclusion of default polyfills
       ${opt('--port')}         port to run the dev server on ${opt('[default: 3000]')}
       ${opt('--reload')}       auto reload the page if hot reloading fails
       ${opt('--title')}        contents for <title> ${opt('[default: React App]')}
@@ -71,7 +70,6 @@ Commands:
     Options:
       ${opt('--force')}        don't shim rendering, use the given entry module directly
       ${opt('--mount-id')}     id for the <div> the app will render into ${opt('[default: app]')}
-      ${opt('--no-polyfill')}  disable bundling of default polyfills
       ${opt('--title')}        contents for <title> ${opt('[default: React App]')}
       ${opt('--vendor')}       create a 'vendor' bundle for node_modules/ modules