+++ src/config/babel.js
@@ -1,37 +1,7 @@
 import {validateOptions} from 'babel-preset-proposals'
 import chalk from 'chalk'
 
-import {padLines, pluralise as s, toSource, typeOf} from '../utils'
-
-// TODO Remove - deprecated
-let warnedAboutStageConfig = false
-const STAGE_3_PROPOSALS = {
-  dynamicImport: true,
-  importMeta: true,
-  classProperties: true,
-  numericSeparator: true,
-  // Actually Stage 4, but still...
-  exportNamespaceFrom: true,
-}
-const STAGE_2_PROPOSALS = {
-  ...STAGE_3_PROPOSALS,
-  decorators: true,
-  functionSent: true,
-  logicalAssignmentOperators: true,
-  throwExpressions: true,
-}
-const STAGE_1_PROPOSALS = {
-  ...STAGE_2_PROPOSALS,
-  exportDefaultFrom: true,
-  pipelineOperator: true,
-  doExpressions: true,
-}
-const PROPOSALS_BY_STAGE = {
-  0: {all: true},
-  1: STAGE_1_PROPOSALS,
-  2: STAGE_2_PROPOSALS,
-  3: STAGE_3_PROPOSALS,
-}
+import {pluralise as s, typeOf} from '../utils'
 
 export function processBabelConfig({report, userConfig}) {
   let {
@@ -45,7 +15,6 @@ export function processBabelConfig({report, userConfig}) {
     react,
     reactConstantElements,
     runtime,
-    stage,
     config,
     ...unexpectedConfig
   } = userConfig.babel
@@ -186,57 +155,6 @@ export function processBabelConfig({report, userConfig}) {
         `Must be an ${chalk.cyan('Object')} (to configure transform-runtime options) or ` +
         `${chalk.cyan('false')} (to disable use of the runtime-transform plugin)`
       )
-    }
-  }
-
-  // TODO Remove - deprecated
-  // stage
-  if ('stage' in userConfig.babel) {
-    let hasProposalsConfig = 'proposals' in userConfig.babel
-    let deprecationMessages = [
-      'Deprecated as of nwb v0.24.0, as Babel 7 no longer has preset-stage-X presets',
-      'Use babel.proposals config instead to enable the required proposal plugins',
-    ]
-
-    if (typeOf(stage) === 'number') {
-      if (stage < 0 || stage > 3) {
-        report.error(
-          'babel.stage',
-          stage,
-          `Must be between ${chalk.cyan(0)} and ${chalk.cyan(3)}`
-        )
-      }
-      else if (!hasProposalsConfig) {
-        deprecationMessages.push(
-          `For now, nwb will enable proposal plugins corresponding to the final version of preset-stage-${stage}`,
-          `This is equivalent to the following babel.proposals config:`,
-          '',
-          ...padLines(toSource(PROPOSALS_BY_STAGE[stage])).split('\n')
-        )
-        userConfig.babel.proposals = {...PROPOSALS_BY_STAGE[stage]}
-      }
-    }
-    else if (stage !== false) {
-      report.error(
-        'babel.stage',
-        stage,
-        `Must be a ${chalk.cyan('Number')} between ${chalk.cyan(0)} and ${chalk.cyan(3)} (to choose a stage preset), ` +
-        `or ${chalk.cyan('false')} (to disable use of a stage preset)`
-      )
-    }
-    else {
-      if (!hasProposalsConfig) {
-        deprecationMessages.push(
-          `For now, nwb will disable use of proposal plugins`,
-          `This is equivalent to configuring ${chalk.cyan('babel.proposals = false')}`
-        )
-        userConfig.babel.proposals = false
-      }
-    }
-
-    if (!warnedAboutStageConfig) {
-      report.deprecated('babel.stage', ...deprecationMessages)
-      warnedAboutStageConfig = true
     }
   }